requirements.txt
python-dotenv
pandas
numpy
matplotlib
pymongo
seaborn
flask
certifi
pymongo[srv]
scikit-learn
pyaml
mlflow
##-e .